Text to doc/requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.4
google-auth==2.23.4
python-dotenv==1.0.0
openai==1.3.7
pytz==2023.3
gunicorn==21.2.0
